chore: clean-up and move components
