+++ packages/design-system/src/components/AppBar/AppBar.tsx
@@ -1,0 +1,16 @@
+/** @jsx jsx */
+import { jsx, Box, BoxProps } from 'theme-ui';
+
+export const AppBar: React.FC<BoxProps> = (props: React.PropsWithChildren<BoxProps>) => (
+  <Box
+    {...props}
+    sx={{
+      bg: 'backgroundInverted',
+      display: 'flex',
+      flexDirection: 'column',
+      alignItems: 'center',
+      justifyContent: 'center',
+    }}
+  />
+);
+export default AppBar;
