+++ packages/design-system/src/components/AppBar/Readme.md
@@ -1,0 +1,5 @@
+### AppBar
+
+```js
+<AppBar/>
+```
