+++ packages/design-system/src/components/AppBar/index.tsx
@@ -1,0 +1,2 @@
+export * from './AppBar';
+export { default } from './AppBar';
