+++ packages/design-system/src/components/AppLayout/Readme.md
@@ -1,0 +1,5 @@
+### AppLayout
+
+```js
+  <AppLayout/>
+```
