+++ packages/design-system/src/components/AppLayout/index.tsx
@@ -1,0 +1,2 @@
+export * from './AppLayout';
+export { default } from './AppLayout';
