+++ packages/design-system/src/components/AppMain/Readme.md
@@ -1,0 +1,5 @@
+### AppMain
+
+```js
+  <AppMain/>
+```
