+++ packages/design-system/src/components/AppMain/index.tsx
@@ -1,0 +1,2 @@
+export * from './AppMain';
+export { default } from './AppMain';
