+++ packages/poolbase-app/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0-semantically-released",
   "description": "",
   "scripts": {
-    "start": "yarn dev",
+    "start": "NODE_ENV=development next dev \"src/app\"",
     "dev": "cd ./src/app && NODE_ENV=development node devServer.js",
     "serve": "NODE_ENV=production next start \"src/app\"",
     "clean": "find ../../dist/app/functions -mindepth 1 ! -regex '^../../dist/app/functions/node_modules/.*' -delete",
@@ -72,6 +72,7 @@
     "puppeteer": "^2.1.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-feather": "^2.0.8",
     "react-firebaseui": "^4.1.0",
     "react-hook-form": "^5.5.1",
     "rxfire": "^3.12.1",
