+++ packages/poolbase-app/src/app/components/Header/index.tsx
@@ -1,0 +1,2 @@
+export * from './Header';
+export { default } from './Header';
