+++ packages/poolbase-app/src/app/components/PageLayout/index.tsx
@@ -1,0 +1,2 @@
+export * from './PageLayout';
+export { default } from './PageLayout';
