+++ packages/poolbase-app/src/app/devServer.js
@@ -8,7 +8,7 @@ const NextI18Next = require('next-i18next').default
 const NextI18NextInstance = new NextI18Next({
   defaultLanguage: 'en',
   otherLanguages: ['de'],
-  localePath: process.env.NODE_ENV === 'production' ? './locales' : '../../locales',
+  localePath: './locales',
   localeSubpaths: {
     de: 'de',
   },
